feat[server/client]:- Book package in user side is done

+++ client/src/services/api.service.ts
@@ -11,14 +11,13 @@ const defaultOptions: ApiOption = {
 
 const handleAPIerror = (error: any, options: ApiOption) => {
   console.error('Api error', error)
-
   if (!options.showToast) return
 
-  const message = error.response.data.message || error.message || 'Request failed';
+  const message = error.data.message || error.response.data.message || error.message || 'Request failed';
   console.log('Api request error', message)
 
-  if (error.response.status == 401) {
-    toast.error("Please login")
+  if (error.status == 401) {
+    toast.error(message)
   } else {
     toast.error(message)
   }
@@ -30,15 +29,17 @@ export const postRequest = async < T = any> (url: string, body: object | FormDat
     if(!(body instanceof FormData)) {
       headers["Content-Type"] = "application/json"
     }
-    const res = await api.post(url,body,{headers})
+    const res = await api.post(url,body,{headers,validateStatus:(status)=> status != 401 && status !== 403})
     console.log('response from api',res)
 
     if(!res.data.success){
-      throw new Error(res.data.message || "Request failed")
+      console.log('res',res)
+      handleAPIerror(res,options)
     }
     return res.data
   } catch (error) {
-    handleAPIerror(error,options)
+    console.log('fucking king')
+    // handleAPIerror(error,options)
     return null
   }
 }
@@ -49,9 +50,9 @@ export const getRequest = async <T = any>(
   options: ApiOption = defaultOptions
 ): Promise<T | null> => {
   try {
-    const res = await api.get(url, params ? { params } : {});
-    if (!res.data.ok) {
-      throw new Error(res.data.message || 'Request failed');
+    const res = await api.get(url, params ? { params , validateStatus:(status)=> status != 401 && status !== 403 } : {validateStatus:(status)=> status != 401 && status !== 403});
+    if (!res.data.success) {
+      handleAPIerror(res,options)
     }
     return res.data;
   } catch (error: any) {
@@ -70,9 +71,9 @@ export const patchRequest = async <T = any>(
     if (!(body instanceof FormData)) {
       headers["Content-Type"] = "application/json";
     }
-    const res = await api.patch(url, body, { headers });
-    if (!res.data.ok) {
-      throw new Error(res.data.message || 'Request failed');
+    const res = await api.patch(url, body, { headers,validateStatus:(status)=> status != 401 && status !== 403 });
+    if (!res.data.success) {
+      handleAPIerror(res,options)
     }
     return res.data;
   } catch (error: any) {
@@ -91,9 +92,9 @@ export const putRequest = async <T = any>(
     if (!(body instanceof FormData)) {
       headers["Content-Type"] = "application/json";
     }
-    const res = await api.put(url, body, { headers });
-    if (!res.data.ok) {
-      throw new Error(res.data.message || 'Request failed');
+    const res = await api.put(url, body, { headers,validateStatus:(status)=> status != 401 && status !== 403 });
+    if (!res.data.success) {
+      handleAPIerror(res,options)
     }
     return res.data;
   } catch (error: any) {
@@ -108,9 +109,9 @@ export const deleteRequest = async <T = any>(
   options: ApiOption = defaultOptions
 ): Promise<T | null> => {
   try {
-    const res = await api.delete(url, params ? { params } : {});
-    if (!res.data.ok) {
-      throw new Error(res.data.message || 'Request failed');
+    const res = await api.delete(url, params ? { params, validateStatus:(status)=> status != 401 && status !== 403 } : {validateStatus:(status)=> status != 401 && status !== 403});
+    if (!res.data.success) {
+      handleAPIerror(res,options)
     }
     return res.data;
   } catch (error: any) {
